Dependency injection. XML serialization instead of Linq2XML.

+++ BooksApp/Global.asax.cs
@@ -1,9 +1,12 @@
-﻿using System.IO;
+﻿using System.Configuration;
+using System.IO;
 using System.Web.Http;
 using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
-using BooksApp.Infrastructure;
+using BooksApp.BLL.Helpers;
+using BooksApp.DAL;
+using BooksApp.Interfaces;
 
 namespace BooksApp
 {
@@ -12,6 +15,16 @@ namespace BooksApp
 
     public class MvcApplication : System.Web.HttpApplication
     {
+        private IBooksHelper BooksHelper { get; set; }
+
+        public MvcApplication()
+        {
+            string connectionString = ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString;
+            var repository = new SqlRepository(connectionString);
+
+            BooksHelper = new BooksHelper(repository);
+        }
+
         protected void Application_Start()
         {
             AreaRegistration.RegisterAllAreas();
@@ -35,7 +48,10 @@ namespace BooksApp
                 if (BooksHelper.ImportBooksFromXmlFile(filePath))
                 {
                     // Either Books were successfully imported into the database, or the xml file can not be parsed correctly. Move the file to Backups.
-                    File.Move(filePath, filePath.Replace(uploadsPath, backupsPath));
+                    if (File.Exists(filePath))
+                    {
+                        File.Move(filePath, filePath.Replace(uploadsPath, backupsPath));
+                    }
                 }
             }
         }
